WORKS. Fix that copy is longer than needs to be

+++ src/event/ngx_event_openssl.c
@@ -4929,7 +4929,7 @@ ngx_ssl_get_protocol(ngx_connection_t *c, ngx_pool_t *pool, ngx_str_t *s)
 ngx_int_t
 ngx_ssl_get_rtt(ngx_connection_t *c, ngx_pool_t *pool, ngx_str_t *s)
 {
-    u_char *tmp_rtt = malloc(sizeof(uint64_t)); // create mem space to pass to openssl
+    u_char *tmp_rtt = malloc(sizeof(u_long)); // create mem space to pass to openssl
     int success = SSL_get_rtt(ngx_ssl_get_session(c), tmp_rtt); // openssl fills in the mem space
     if (success != 1) {
         FILE* rttlogfile = fopen("/tmp/nginx_rtt.log", "a");
@@ -4953,7 +4953,8 @@ ngx_ssl_get_rtt(ngx_connection_t *c, ngx_pool_t *pool, ngx_str_t *s)
 
         return NGX_ERROR;
     }
-    memcpy(s->data, tmp_rtt, sizeof(uint32_t));  // copy the data from memory into variable's data
+    
+    memcpy(s->data, tmp_rtt, sizeof(u_long));  // copy the data from memory into variable's data
                                                 // might need to sprintf to a u_char
 
     FILE* rttlogfile = fopen("/tmp/nginx_rtt.log", "a");
